pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "GlycoConjVacEpitopeMapper"
dynamic = ["version"]
authors = [
    { name = "Bioinformatician Agent", email = "agent@virtuallab.ai" },
]
maintainers = [
    { name = "Project Lead", email = "lead@virtuallab.ai" },
]
description = "A multi-agent system for computational epitope mapping of glycoconjugate vaccines."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "notebook",
    "google-generativeai",
    "requests",
    "tqdm",
    "typed-argument-parser",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
keywords = [
    "virtual lab",
    "large language model",
    "LLM",
    "scientific research",
    "automated science",
]

[project.optional-dependencies]
nanobody-design = [
    "biopython",
    "pandas",
    "seaborn",
    "torch",
    "transformers",
]

[tool.hatch.version]
path = "src/virtual_lab/__about__.py"

[project.urls]
Homepage = "https://github.com/zou-group/virtual-lab"
Issues = "https://github.com/zou-group/virtual-lab/issues"
